+++ src/azure-cli/azure/cli/command_modules/acs/_format.py
@@ -10,6 +10,31 @@ from jmespath import compile as compile_jmes, Options
 from jmespath import functions
 
 
+def aks_agentpool_show_table_format(result):
+    """Format an agent pool as summary results for display with "-o table"."""
+    return [_aks_agentpool_table_format(result)]
+
+
+def _aks_agentpool_table_format(result):
+    parsed = compile_jmes("""{
+        name: name,
+        osType: osType,
+        kubernetesVersion: kubernetesVersion,
+        vmSize: vmSize,
+        osDiskSizeGB: osDiskSizeGB,
+        count: count,
+        maxPods: maxPods,
+        provisioningState: provisioningState
+    }""")
+    # use ordered dicts so headers are predictable
+    return parsed.search(result, Options(dict_cls=OrderedDict))
+
+
+def aks_agentpool_list_table_format(results):
+    """Format an agent pool list for display with "-o table"."""
+    return [_aks_agentpool_table_format(r) for r in results]
+
+
 def aks_list_table_format(results):
     """"Format a list of managed clusters as summary results for display with "-o table"."""
     return [_aks_table_format(r) for r in results]