+++ src/azure-cli/azure/cli/command_modules/appservice/commands.py
@@ -8,7 +8,7 @@ from azure.cli.core.commands import CliCommandType
 from azure.cli.core.util import empty_on_404
 
 from ._client_factory import cf_web_client, cf_plans, cf_webapps
-from ._validators import validate_app_exists_in_rg, validate_app_or_slot_exists_in_rg
+from ._validators import validate_app_exists_in_rg, validate_app_or_slot_exists_in_rg, validate_asp_exists_in_rg
 
 
 def output_slots_in_table(slots):
@@ -166,7 +166,7 @@ def load_command_table(self, _):
 
     with self.command_group('webapp deployment source') as g:
         g.custom_command('config-local-git', 'enable_local_git')
-        g.custom_command('config-zip', 'enable_zip_deploy')
+        g.custom_command('config-zip', 'enable_zip_deploy_webapp')
         g.custom_command('config', 'config_source_control', exception_handler=ex_handler_factory())
         g.custom_command('sync', 'sync_site_repo', exception_handler=ex_handler_factory())
         g.custom_show_command('show', 'show_source_control')
@@ -234,7 +234,9 @@ def load_command_table(self, _):
         g.command('delete', 'delete', confirmation=True)
         g.custom_command('list', 'list_app_service_plans')
         g.show_command('show', 'get')
-        g.generic_update_command('update', custom_func_name='update_app_service_plan', setter_arg_name='app_service_plan')
+        g.generic_update_command('update', custom_func_name='update_app_service_plan', setter_arg_name='app_service_plan',
+                                 validator=validate_asp_exists_in_rg)
+
     with self.command_group('appservice') as g:
         g.custom_command('list-locations', 'list_locations', transform=transform_list_location_output)