+++ src/azure-cli/azure/cli/command_modules/vm/_client_factory.py
@@ -108,3 +108,13 @@ def cf_dedicated_hosts(cli_ctx, _):
 
 def cf_dedicated_host_groups(cli_ctx, _):
     return _compute_client_factory(cli_ctx).dedicated_host_groups
+
+
+def _log_analytics_client_factory(cli_ctx, subscription_id, *_):
+    from azure.mgmt.loganalytics import LogAnalyticsManagementClient
+    from azure.cli.core.commands.client_factory import get_mgmt_service_client
+    return get_mgmt_service_client(cli_ctx, LogAnalyticsManagementClient, subscription_id=subscription_id)
+
+
+def cf_log_analytics(cli_ctx, subscription_id, *_):
+    return _log_analytics_client_factory(cli_ctx, subscription_id).workspaces